+++ crates/starknet-replay/src/block_context.rs
@@ -1,0 +1,45 @@
+use blockifier::block::{BlockInfo, GasPrices};
+use blockifier::context::{BlockContext, ChainInfo, FeeTokenAddresses};
+use blockifier::versioned_constants::VersionedConstants;
+use starknet::core::types::BlockWithTxs;
+use starknet_api::block::{BlockNumber, BlockTimestamp};
+use starknet_api::core::{ContractAddress, PatriciaKey};
+use starknet_api::hash::StarkHash;
+use starknet_api::{contract_address, patricia_key};
+
+use crate::utils::felt_to_u128;
+
+pub fn build_block_context(chain_id: String, block: &BlockWithTxs) -> BlockContext {
+    let sequencer_address_hex = block.sequencer_address.to_hex_string();
+    let sequencer_address = contract_address!(sequencer_address_hex.as_str());
+
+    let block_info = BlockInfo {
+        block_number: BlockNumber(block.block_number),
+        block_timestamp: BlockTimestamp(block.timestamp),
+        sequencer_address,
+        gas_prices: GasPrices {
+            eth_l1_gas_price: felt_to_u128(&block.l1_gas_price.price_in_wei).try_into().unwrap(),
+            strk_l1_gas_price: felt_to_u128(&block.l1_gas_price.price_in_fri).try_into().unwrap(),
+            eth_l1_data_gas_price: felt_to_u128(&block.l1_data_gas_price.price_in_wei).try_into().unwrap(),
+            strk_l1_data_gas_price: felt_to_u128(&block.l1_data_gas_price.price_in_fri).try_into().unwrap(),
+        },
+        use_kzg_da: false,
+    };
+
+    let chain_info = ChainInfo {
+        chain_id: starknet_api::core::ChainId(chain_id),
+        // cf. https://docs.starknet.io/tools/important-addresses/
+        fee_token_addresses: FeeTokenAddresses {
+            strk_fee_token_address: contract_address!(
+                "0x04718f5a0fc34cc1af16a1cdee98ffb20c31f5cd61d6ab07201858f4287c938d"
+            ),
+            eth_fee_token_address: contract_address!(
+                "0x049d36570d4e46f48e99674bd3fcc84644ddd6b96f7c741b1562b82f9e004dc7"
+            ),
+        },
+    };
+
+    let versioned_constants = VersionedConstants::latest_constants();
+
+    BlockContext::new_unchecked(&block_info, &chain_info, versioned_constants)
+}
