+++ crates/starknet-replay/src/lib.rs
@@ -1,2 +1,3 @@
+pub mod block_context;
 pub mod rpc_state_reader;
 pub(crate) mod utils;
